+++ test/test.js
@@ -1,8 +1,10 @@
-var EventEmitter = require('events').EventEmitter
 
-var http = require('http')
-var onFinished = require('..')
+var EventEmitter = require('events').EventEmitter
 var should = require('should')
+var assert = require('assert')
+var http = require('http')
+
+var onFinished = require('..')
 
 function createThingie() {
   var ee = new EventEmitter
@@ -23,22 +25,14 @@ describe('on socket error', function () {
     called.should.be.true
   })
 
-  it('should not execute the callback if response is finished', function () {
+  it('should not execute the callback if response is finished', function (done) {
     var thingie = createThingie()
-    var called = false
-    onFinished(thingie, function () {
-      called = true
+    onFinished(thingie, function (err) {
+      assert.ifError(err)
+      done()
     })
     thingie.emit('finish')
-    try {
-      // throws if there are no listeners
-      thingie.socket.emit('error', new Error('boom'))
-      throw new Error('alksdjflaksjdf')
-    } catch (err) {
-      if (err.message !== 'boom')
-        throw new Error('wtf')
-    }
-    called.should.be.true
+    thingie.socket.emit(new Error('boom'))
   })
 })
 
@@ -116,3 +110,43 @@ describe('http', function () {
     })
   })
 })
+
+describe('event emitter leaks', function () {
+  describe('when adding a lot of listeners on the same request', function () {
+    it('should not warn and add at most 1 listener per emitter per event', function () {
+      // we just have to make sure tests pass without a bunch of logs
+      var thingie = createThingie()
+      var called = false
+
+      onFinished(thingie, function (err) {
+        called = true
+        err.message.should.equal('boom')
+      })
+
+      for (var i = 0; i < 1000; i++) {
+        onFinished(thingie, noop)
+      }
+
+      assert.equal(1, thingie.socket.listeners('error').length)
+      assert.equal(1, thingie.socket.listeners('close').length)
+      assert.equal(1, thingie.listeners('finish').length)
+
+      thingie.socket.emit('error', new Error('boom'))
+      called.should.be.true
+    })
+  })
+
+  it('should clean up after itself', function (done) {
+    var thingie = createThingie()
+    onFinished(thingie, function () {
+      assert(!thingie.socket.listeners('error').length)
+      assert(!thingie.socket.listeners('close').length)
+      assert(!thingie.listeners('finish').length)
+      done()
+    })
+
+    thingie.socket.emit('error')
+  })
+})
+
+function noop() {}